Adds tests.
