+++ Packages/ChineseNumbers/Tests/ChineseNumbersTests/ChineseNumbersTests.swift
@@ -47,47 +47,18 @@ final class ChineseNumbersTests: XCTestCase {
         XCTAssertEqual(output, "一億〇一")
     }
 
-    func testSuzhou1() {
-        let intInput = "1000"
-        let decInput = "0000"
-        let output = SuzhouNumbers.generate(intPart: intInput, decPart: decInput)
-        XCTAssertEqual(output, "〡\n千")
+    func testNum4() {
+        let intInput = "1"
+        let decInput = "99999"
+        let output = ChineseNumbers.generate(intPart: intInput, decPart: decInput, digitCase: .lowercase)
+        XCTAssertEqual(output, "一點九九九九九")
     }
 
-    func testSuzhou2() {
-        let intInput = "1234"
-        let decInput = "0000"
-        let output = SuzhouNumbers.generate(intPart: intInput, decPart: decInput)
-        XCTAssertEqual(output, "〡二〣〤\n千")
+    func testNum5() {
+        let intInput = "1"
+        let decInput = "99999"
+        let output = ChineseNumbers.generate(intPart: intInput, decPart: decInput, digitCase: .uppercase)
+        XCTAssertEqual(output, "壹點玖玖玖玖玖")
     }
-
-    func testSuzhou3() {
-        let intInput = "2222"
-        let decInput = "0000"
-        let output = SuzhouNumbers.generate(intPart: intInput, decPart: decInput)
-        XCTAssertEqual(output, "〢二〢二\n千")
-    }
-
-    func testSuzhou4() {
-        let intInput = "0022"
-        let decInput = "2200"
-        let output = SuzhouNumbers.generate(intPart: intInput, decPart: decInput)
-        XCTAssertEqual(output, "〢二〢二\n十")
-    }
-
-    func testSuzhou5() {
-        let intInput = "0022"
-        let decInput = "2200"
-        let output = SuzhouNumbers.generate(intPart: intInput, decPart: decInput, preferInitialVertical: false)
-        XCTAssertEqual(output, "二〢二〢\n十")
-    }
-
-    func testSuzhou6() {
-        let intInput = "0010"
-        let decInput = "0000"
-        let output = SuzhouNumbers.generate(intPart: intInput, decPart: decInput, preferInitialVertical: false)
-        XCTAssertEqual(output, "〸")
-    }
-
 
 }
